REFACTOR split enrichment logic into pieces
